apps/api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.3
plotly==5.22.0
python-dotenv==1.0.1
smolagents==1.9.2
-e ../../packages/agents 